nix inline svg
